Fix failures when using npipe monitoring endpoints (#1371)
(cherry picked from commit e184051717ba412a78e75994e647738229120365)

+++ internal/pkg/core/monitoring/beats/sidecar_monitor.go
@@ -88,6 +88,10 @@ func (b *SidecarMonitor) EnrichArgs(spec program.Spec, pipelineID string, args [
 func (b *SidecarMonitor) Cleanup(spec program.Spec, pipelineID string) error {
 	endpoint := MonitoringEndpoint(spec, b.operatingSystem, pipelineID, true)
 	drop := monitoringDrop(endpoint)
+	if drop == "" {
+		// not exposed using sockets
+		return nil
+	}
 
 	return os.RemoveAll(drop)
 }
@@ -103,6 +107,11 @@ func (b *SidecarMonitor) Close() {
 func (b *SidecarMonitor) Prepare(spec program.Spec, pipelineID string, uid, gid int) error {
 	endpoint := MonitoringEndpoint(spec, b.operatingSystem, pipelineID, true)
 	drop := monitoringDrop(endpoint)
+
+	if drop == "" {
+		// not exposed using sockets
+		return nil
+	}
 
 	if err := os.MkdirAll(drop, 0775); err != nil {
 		return errors.New(err, fmt.Sprintf("failed to create a directory %q", drop))
